+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 (Please put an entry here in each PR)
 
 - Use ```$AWS_PROFILE``` to obtain default profile name 
+- Implement `Default` for `Region`
 
 ## [0.31.0] - 2018-01-21